# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQ Test Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateResultBody {
  /** Display name or location */
  userName?: string;
  /** IQ score (typically 70-145) */
  score: number;
  /** Time in seconds */
  timeTaken: number;
  correctAnswers: number;
  totalQuestions: number;
}

export interface TestResult {
  id: number;
  userName?: string | null;
  score: number;
  timeTaken: number;
  correctAnswers: number;
  totalQuestions: number;
  percentile: number;
  createdAt: string;
}

export interface LeaderboardEntry {
  id: number;
  userName?: string | null;
  score: number;
  timeTaken: number;
  createdAt: string;
}

export interface TestStats {
  totalTests: number;
  averageScore: number;
  highestScore: number;
  testsToday: number;
}

export interface ErrorResponse {
  error: string;
}