# lib/api-zod/src/generated/types/createResultBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQ Test Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateResultBody {
  /** Display name or location */
  userName?: string;
  /** IQ score (typically 70-145) */
  score: number;
  /** Time in seconds */
  timeTaken: number;
  correctAnswers: number;
  totalQuestions: number;
}