# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQ Test Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  id: number;
  userName?: string | null;
  score: number;
  timeTaken: number;
  createdAt: Date;
}