# lib/api-zod/src/generated/types/testStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IQ Test Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TestStats {
  totalTests: number;
  averageScore: number;
  highestScore: number;
  testsToday: number;
}